Rename constant 'AI state' to 'AI state bandit' on server, and fix related code parts in Instance and Patrol classes on server
