+++ nodejs-server/src/modules/patrols/AIStateBandit.js
